refactor: [UF-361] IconWrapper에 cva variants 적용
