--- src/shared/ui/Icons/IconWrapper.styles.ts
@@ -1,10 +1,0 @@
-export const spanStyleMap = {
-  base: 'inline-flex items-center justify-center shrink-0',
-} as const;
-
-export const defaultValues = {
-  size: 'md',
-  color: 'current',
-  className: '',
-  children: null,
-} as const;
